refactor: updated borrow and return APIs

+++ models/Book.js
@@ -1,14 +1,26 @@
 const { dbConnection } = require("../config/db");
 const { ObjectId } = require("mongodb");
 
-const findAllBook = async () => {
+const connectToBookCollection = async () => {
   const db = await dbConnection();
-  const bookCollection = db.collection("books");
-  const allBook = bookCollection.find({}).toArray();
+  return db.collection("books");
+};
+
+const connectToBorrowedCollection = async () => {
+  const db = await dbConnection();
+  return db.collection("borrowed");
+};
+
+const findAllBook = async () => {
+  const bookCollection = await connectToBookCollection();
+  const allBook = await bookCollection.find({}).toArray();
   return allBook;
 };
 
 const removeBook = async (bookId) => {
+  if (!ObjectId.isValid(bookId)) {
+    throw new Error("Invalid bookId format");
+  }
   const db = await dbConnection();
   const bookCollection = db.collection("books");
   const result = await bookCollection.deleteOne({ _id: new ObjectId(bookId) });
@@ -20,10 +32,53 @@ const removeBook = async (bookId) => {
 };
 
 const addBook = async (newBook) => {
-  const db = await dbConnection();
-  const bookCollection = db.collection("books");
+  const bookCollection = await connectToBookCollection();
   try {
-    return await bookCollection.insertOne(newBook);
+    await bookCollection.insertOne(newBook);
+    return { massage: "New Book is Added Successfully" };
+  } catch (e) {
+    return e;
+  }
+};
+
+const returnBook = async (bookId) => {
+  console.log(bookId, "bookId");
+  try {
+    const bookCollection = await connectToBookCollection();
+    const borrowedCollection = await connectToBorrowedCollection();
+    await bookCollection.findOneAndUpdate(
+      { _id: new ObjectId(bookId) },
+      { $set: { free: 1 } }
+    );
+    await borrowedCollection.deleteOne({ bookId: bookId });
+    return { message: "Book updated successfully" };
+  } catch (e) {
+    return e;
+  }
+};
+
+const borrowBook = async (userId, bookId) => {
+  try {
+    const borrowedCollection = await connectToBorrowedCollection();
+    const bookCollection = await connectToBookCollection();
+    const isAvailableBook = await bookCollection.findOne({
+      _id: new ObjectId(bookId),
+      free: 1,
+    });
+
+    if (isAvailableBook) {
+      await borrowedCollection.insertOne({
+        userId,
+        bookId,
+      });
+      await bookCollection.updateOne(
+        { _id: new ObjectId(bookId) },
+        { $set: { free: 0 } }
+      );
+      return { message: "Book successfully borrowed" };
+    } else {
+      return { message: "Book is not available for borrowing" };
+    }
   } catch (e) {
     return e;
   }
@@ -33,4 +88,6 @@ module.exports = {
   findAllBook,
   removeBook,
   addBook,
+  returnBook,
+  borrowBook,
 };
